+++ src/app/(auth)/authService.ts
@@ -2,7 +2,7 @@
 
 import { cookies } from "next/headers";
 import { handleAsync } from "@/utils/handleAsync";
-import { type AsyncActionResponse } from "@/modules/shared/types";
+import { type AsyncActionResponse } from "@/utils/handleAsync";
 import { getAccessToken, getRefreshToken } from "@/utils/getCookie";
 import { POST, UNAUTHPOST } from "@/utils/requests";
