+++ src/app/(main)/my-voyage/[teamId]/directory/components/DirectoryComponentWrapper.tsx
@@ -10,7 +10,7 @@ import { type TeamDirectory } from "@/store/features/directory/directorySlice";
 
 import { getAccessToken } from "@/utils/getCookie";
 import { handleAsync } from "@/utils/handleAsync";
-import { type AsyncActionResponse } from "@/modules/shared/types";
+import { type AsyncActionResponse } from "@/utils/handleAsync";
 import { GET } from "@/utils/requests";
 import { CacheTag } from "@/utils/cacheTag";
 import { type User } from "@/store/features/user/userSlice";
