+++ src/app/(main)/my-voyage/[teamId]/directory/directoryService.ts
@@ -3,7 +3,7 @@
 import { revalidateTag } from "next/cache";
 import { getAccessToken } from "@/utils/getCookie";
 import { handleAsync } from "@/utils/handleAsync";
-import { type AsyncActionResponse } from "@/modules/shared/types";
+import { type AsyncActionResponse } from "@/utils/handleAsync";
 import { PATCH } from "@/utils/requests";
 import { CacheTag } from "@/utils/cacheTag";
