+++ src/app/(main)/my-voyage/[teamId]/features/featuresService.ts
@@ -5,7 +5,7 @@ import { type Features } from "@/store/features/features/featuresSlice";
 import { CacheTag } from "@/utils/cacheTag";
 import { getAccessToken } from "@/utils/getCookie";
 import { handleAsync } from "@/utils/handleAsync";
-import { type AsyncActionResponse } from "@/modules/shared/types";
+import { type AsyncActionResponse } from "@/utils/handleAsync";
 import { DELETE, PATCH, POST } from "@/utils/requests";
 
 interface SaveOrderProps {
