+++ src/app/(main)/my-voyage/[teamId]/sprints/sprintsService.ts
@@ -4,7 +4,7 @@ import { revalidateTag } from "next/cache";
 import { getAccessToken } from "@/utils/getCookie";
 import { DELETE, PATCH, POST } from "@/utils/requests";
 import { handleAsync } from "@/utils/handleAsync";
-import { type AsyncActionResponse } from "@/modules/shared/types";
+import { type AsyncActionResponse } from "@/utils/handleAsync";
 import { CacheTag } from "@/utils/cacheTag";
 import { getSprintCache } from "@/utils/getSprintCache";
