+++ src/app/(main)/my-voyage/[teamId]/tech-stack/techStackService.ts
@@ -5,7 +5,7 @@ import type { Category } from "./finalize/types";
 import { CacheTag } from "@/utils/cacheTag";
 import { getAccessToken } from "@/utils/getCookie";
 import { handleAsync } from "@/utils/handleAsync";
-import { type AsyncActionResponse } from "@/modules/shared/types";
+import { type AsyncActionResponse } from "@/utils/handleAsync";
 import { DELETE, PATCH, POST } from "@/utils/requests";
 import { type TechStackItem } from "@/store/features/techStack/techStackSlice";
