--- src/modules/shared/types.ts
@@ -1,6 +1,0 @@
-export interface AppError {
-  message: string;
-}
-
-export type AsyncFunction<X> = () => Promise<X>;
-export type AsyncActionResponse<X> = [X | null, AppError | null];
