+++ src/types/types.ts
@@ -1,1 +1,4 @@
 // only put global types here
+export interface AppError {
+  message: string;
+}
